+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Handle Impure Decoder
   - Cache response before Impure Decoder decodes the response
   - Fix scope of the fetch query/send command methods, which may overlaps with nested queries in impure decoding
+- Allow setting verbosity for a profiler
 
 ## 1.2.1 - 2022-03-09
 - Fix profiling last used decoders when there are none